fix(api): send the invite's permissions as a number
discord.QueryValues formats values with fmt, and Permissions.String()
lists permission names, so the invite link carried
"permissions=Send Messages in Threads, ..." instead of the bitfield
Discord expects. The test now checks for the exact number.

+++ internal/api/server.go
@@ -98,7 +98,8 @@ func (s *Server) invite(w http.ResponseWriter, r *http.Request) {
 	values := discord.QueryValues{
 		"client_id":   s.cfg.DiscordClientID,
 		"scope":       "bot applications.commands",
-		"permissions": config.BotPermissions,
+		// Discord wants the bitfield as a number; Permissions' String() lists names.
+		"permissions": int64(config.BotPermissions),
 	}
 	if id, err := snowflake.Parse(r.URL.Query().Get("guild_id")); err == nil {
 		values["guild_id"] = id
